requirements.txt
asgiref==3.7.2
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.2.0
Django==4.2.2
django-crispy-forms==2.0
django-mssql-backend==2.8.1
django-pyodbc==1.1.3
django-widget-tweaks==1.4.12
djangorestframework==3.14.0
drf-yasg==1.21.7
et-xmlfile==1.1.0
idna==3.4
inflection==0.5.1
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
mssql-django==1.3
openpyxl==3.1.2
packaging==23.1
Pillow==10.0.0
psycopg2==2.9.6
pyodbc==4.0.39
pytz==2023.3
PyYAML==6.0.1
referencing==0.30.2
requests==2.31.0
rpds-py==0.10.3
sqlparse==0.4.4
style==1.1.0
swagger-spec-validator==3.0.3
typing_extensions==4.7.0
tzdata==2023.3
update==0.0.1
uritemplate==4.1.1
urllib3==2.0.4
